pyproject.toml
[project]
name = "chimera-factory"
version = "0.1.0"
description = "Project Chimera: The Factory that builds Autonomous AI Influencers"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chimera_factory"]

[tool.uv]
# Note: dev-dependencies is deprecated but still works
# The warning can be ignored for now - dependency-groups format is still evolving
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
# Pytest configuration for Project Chimera
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Test markers for different test types
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (require external services)",
    "contracts: Contract tests (API/Skill validation)",
    "e2e: End-to-end tests (full system)",
    "slow: Slow running tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/chimera_factory"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
